requirements.txt
async-timeout==4.0.2
autopep8==1.6.0
celery==5.3.0
colorlog==6.9.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
Deprecated==1.2.13
Django==3.2.19
django-celery-beat==2.2.1
django-celery-results==2.5.1
django-ckeditor==6.5.0
django-cors-headers==3.13.0
django-crontab==0.7.1
django-extensions==3.1.5
django-filter==21.1
django-js-asset==2.0.0
django-oauth-toolkit==2.0.0
django-oauth2-provider==0.2.6.1
django-redis==5.4.0
django-ses==3.3.0
django-storages==1.12.3
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
drf-errors-formatter==1.0.1
drf-yasg==1.20.0
Jinja2==3.1.2
kombu==5.3.7
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==1.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.1
redis==4.3.3
requests==2.28.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.27.0
